Change nativeSize from Int to Long for FFM compatibility
Java's Foreign Function & Memory (FFM) API uses long for memory segment
sizes. Changing NativeMemoryAccess.nativeSize to Long enables seamless
interop with FFM MemorySegments without truncation concerns.

Updated all implementations:
- DirectJvmBuffer (JVM and Android)
- MutableDataBuffer and NSDataBuffer (Apple)
- LinearBuffer (WASM)
- JsBuffer (JavaScript)

+++ buffer/src/androidMain/kotlin/com/ditchoom/buffer/DirectJvmBuffer.kt
@@ -28,7 +28,7 @@ class DirectJvmBuffer(
     /**
      * The size of the native memory region in bytes.
      */
-    override val nativeSize: Int get() = capacity
+    override val nativeSize: Long get() = capacity.toLong()
 
     companion object {
         private val addressField: Field? by lazy {
